refactor: share the pagination decision instead of copying it per adapter
`PaginationStep`, `advance_cursor` and `pagination_step` were duplicated
verbatim across the Anthropic and Gemini adapters — identical logic, identical
doc comments. That is the defect this repo keeps paying for: a rule written
twice, where hardening one copy leaves the other behind and nothing fails.
Notably it arrived in the very commit that fixed a truncation bug in one
branch of that loop while two other branches still returned a partial list as
success.

All three now live once in `ai_provider/mod.rs`, generic over the cursor type,
with the page budget passed in rather than read from a module-private const.
No "keep in sync" comment is left behind — that comment is what has been
lying.

The loop also had no integration coverage. Its pure decision was tested, but
none of the wiring was: that a cursor actually reaches the next request's
query, that a mid-pagination failure rejects the whole fetch instead of
returning the pages already collected, that a malformed later page errors, and
that the cumulative deadline fires at all. A deadline that silently never
triggers is exactly the sort of guarantee this branch keeps finding was never
true.

Both adapters now expose a `list_models_transport` seam taking `base` and
`total_deadline` — the same split `OpenAiClient` already used — so wiremock
can drive all four cases per adapter. The deadline test runs a 30ms budget
against a 150ms delayed response, so it proves the bound in ~40ms instead of
waiting on the real one.

Moving the shared code into `mod.rs` pushed it to 1410 lines, past the R8 cap,
so its test module moves to a `tests.rs` sibling — the same split
`ai_config/mod.rs` already uses. `mod.rs` is now 1076.
