+++ apps/desktop/src-tauri/src/commands/ai_provider/gemini.rs
@@ -13,9 +13,9 @@ use super::retry::send_with_retry;
 use super::stream::{stream_response, StreamPiece};
 use super::timeouts;
 use super::{
-    friendly_api_error, model_entry, single_shot_turn, split_system, AgentTurn, AiGenerateRequest,
-    AiProvider, ChatMsg, ModelCapabilities, ProviderId, RequestTrace, Role, StopReason, TokenParam,
-    ToolCall, ToolSpec, Usage,
+    friendly_api_error, model_entry, pagination_step, single_shot_turn, split_system, AgentTurn,
+    AiGenerateRequest, AiProvider, ChatMsg, ModelCapabilities, PaginationStep, ProviderId,
+    RequestTrace, Role, StopReason, TokenParam, ToolCall, ToolSpec, Usage,
 };
 
 const BASE: &str = "https://generativelanguage.googleapis.com";
@@ -94,52 +94,6 @@ fn parse_model_page(body: &Value) -> AppResult<(Vec<Value>, Option<String>)> {
         .filter(|t| !t.is_empty())
         .map(String::from);
     Ok((names, next_page_token))
-}
-
-/// Whether pagination should continue to `next` — `None` when there's no next
-/// page OR when `next` doesn't differ from `current` (a provider returning a
-/// non-advancing token must stop the loop, not re-fetch the same page up to
-/// `MAX_LIST_MODELS_PAGES` times, duplicating entries). Pure so the
-/// progress-guard is unit-testable without a network mock.
-fn advance_cursor(current: &Option<String>, next: Option<String>) -> Option<String> {
-    if next.is_none() || &next == current {
-        None
-    } else {
-        next
-    }
-}
-
-/// Outcome of one `list_models` pagination iteration — see [`pagination_step`].
-#[derive(Debug, Clone, PartialEq, Eq)]
-enum PaginationStep {
-    /// Fetch another page with this token.
-    Continue(String),
-    /// A genuine stopping point: no next page, or [`advance_cursor`]'s
-    /// progress guard caught a stuck token. `Ok(all)`.
-    Done,
-    /// Ran out of `MAX_LIST_MODELS_PAGES` while the token was STILL
-    /// genuinely advancing — there IS more catalogue this fetch didn't
-    /// cover. Must reject rather than silently return an incomplete list.
-    Incomplete,
-}
-
-/// One step of the `MAX_LIST_MODELS_PAGES`-bounded pagination loop's control
-/// flow: given the 0-based index of the page JUST fetched and the raw `next`
-/// token it reported, decide whether to continue, stop cleanly, or stop
-/// incomplete. Pure (no I/O) so the exact `MAX_LIST_MODELS_PAGES` boundary is
-/// unit-testable without 50 live HTTP round-trips — `list_models`'s loop
-/// calls this once per page and dispatches on the result, so this function
-/// (not a hand-duplicated copy) is what actually runs in production.
-fn pagination_step(
-    page_index: usize,
-    current: &Option<String>,
-    next: Option<String>,
-) -> PaginationStep {
-    match advance_cursor(current, next) {
-        Some(id) if page_index + 1 < MAX_LIST_MODELS_PAGES => PaginationStep::Continue(id),
-        Some(_) => PaginationStep::Incomplete,
-        None => PaginationStep::Done,
-    }
 }
 
 /// Concatenate every `parts[].text` of the first candidate (non-streaming
@@ -838,6 +792,78 @@ impl GeminiClient {
         trace.end(Some(status.as_u16()), true);
         Ok(join_parts_text(&data))
     }
+
+    /// The full paginated `/v1beta/models` transport — no `AppHandle`, so
+    /// it's directly testable against a `wiremock::MockServer` by passing
+    /// its `uri()` as `base` (production always calls this with `BASE`).
+    /// Mirrors [`OpenAiClient::list_models_transport`](super::openai::OpenAiClient::list_models_transport).
+    ///
+    /// Loops through `pageToken` pages up to `MAX_LIST_MODELS_PAGES`,
+    /// bounded by a SINGLE cumulative `total_deadline` across every page
+    /// (not a fresh timeout per request) — also a parameter (production
+    /// always passes `timeouts::LIST_MODELS_TOTAL`) so a test can force it
+    /// to expire in milliseconds instead of 30 real seconds.
+    async fn list_models_transport(
+        &self,
+        base: &str,
+        api_key: &str,
+        total_deadline: std::time::Duration,
+    ) -> AppResult<Vec<Value>> {
+        let client = crate::net::http::shared();
+        let mut all = Vec::new();
+        let mut page_token: Option<String> = None;
+        let deadline = tokio::time::Instant::now() + total_deadline;
+        for page_index in 0..MAX_LIST_MODELS_PAGES {
+            let mut req = client
+                .get(format!("{base}/v1beta/models"))
+                .header("x-goog-api-key", api_key)
+                // No explicit `pageSize` — see the matching note in
+                // `anthropic.rs`: the `pageToken` loop already yields every
+                // page, so an unverified page-size ceiling would risk 400-ing
+                // the whole listing to save one cached round-trip.
+                .timeout(timeouts::LIST_MODELS);
+            if let Some(token) = &page_token {
+                req = req.query(&[("pageToken", token.as_str())]);
+            }
+            let resp = match tokio::time::timeout_at(deadline, req.send()).await {
+                Ok(Ok(r)) => r,
+                Ok(Err(e)) => {
+                    return Err(AppError::Network(format!(
+                        "{}: request failed: {e}",
+                        self.id().as_str()
+                    )))
+                }
+                Err(_elapsed) => {
+                    return Err(AppError::Network(format!(
+                        "{}: timed out listing models across multiple pages",
+                        self.id().as_str()
+                    )))
+                }
+            };
+            let status = resp.status();
+            if !status.is_success() {
+                let body_text = resp.text().await.unwrap_or_default();
+                return Err(friendly_api_error(self.id(), status, &body_text));
+            }
+            let body: Value = resp
+                .json()
+                .await
+                .map_err(|e| AppError::Provider(format!("{}: parse: {e}", self.id().as_str())))?;
+            let (mut page, next) = parse_model_page(&body)?;
+            all.append(&mut page);
+            match pagination_step(page_index, MAX_LIST_MODELS_PAGES, &page_token, next) {
+                PaginationStep::Continue(token) => page_token = Some(token),
+                PaginationStep::Done => break,
+                PaginationStep::Incomplete => {
+                    return Err(AppError::Provider(format!(
+                        "{}: model catalogue has more than {MAX_LIST_MODELS_PAGES} pages — stopped early rather than return an incomplete list",
+                        self.id().as_str()
+                    )))
+                }
+            }
+        }
+        Ok(all)
+    }
 }
 
 #[async_trait]
@@ -1056,63 +1082,8 @@ impl AiProvider for GeminiClient {
         // Pro-tier default in `provider-meta.ts` is a `-preview` id, which
         // would silently vanish from the picker on `/v1`).
         let api_key = require_gemini_key(app)?;
-        let client = crate::net::http::shared();
-        let mut all = Vec::new();
-        let mut page_token: Option<String> = None;
-        // One deadline for the WHOLE paginated fetch, not per-request — a
-        // per-request-only timeout lets up to `MAX_LIST_MODELS_PAGES` individual
-        // `LIST_MODELS` timeouts chain into one very long invoke.
-        let deadline = tokio::time::Instant::now() + timeouts::LIST_MODELS_TOTAL;
-        for page_index in 0..MAX_LIST_MODELS_PAGES {
-            let mut req = client
-                .get(format!("{BASE}/v1beta/models"))
-                .header("x-goog-api-key", &api_key)
-                // No explicit `pageSize` — see the matching note in
-                // `anthropic.rs`: the `pageToken` loop already yields every
-                // page, so an unverified page-size ceiling would risk 400-ing
-                // the whole listing to save one cached round-trip.
-                .timeout(timeouts::LIST_MODELS);
-            if let Some(token) = &page_token {
-                req = req.query(&[("pageToken", token.as_str())]);
-            }
-            let resp = match tokio::time::timeout_at(deadline, req.send()).await {
-                Ok(Ok(r)) => r,
-                Ok(Err(e)) => {
-                    return Err(AppError::Network(format!(
-                        "{}: request failed: {e}",
-                        self.id().as_str()
-                    )))
-                }
-                Err(_elapsed) => {
-                    return Err(AppError::Network(format!(
-                        "{}: timed out listing models across multiple pages",
-                        self.id().as_str()
-                    )))
-                }
-            };
-            let status = resp.status();
-            if !status.is_success() {
-                let body_text = resp.text().await.unwrap_or_default();
-                return Err(friendly_api_error(self.id(), status, &body_text));
-            }
-            let body: Value = resp
-                .json()
-                .await
-                .map_err(|e| AppError::Provider(format!("{}: parse: {e}", self.id().as_str())))?;
-            let (mut page, next) = parse_model_page(&body)?;
-            all.append(&mut page);
-            match pagination_step(page_index, &page_token, next) {
-                PaginationStep::Continue(token) => page_token = Some(token),
-                PaginationStep::Done => break,
-                PaginationStep::Incomplete => {
-                    return Err(AppError::Provider(format!(
-                        "{}: model catalogue has more than {MAX_LIST_MODELS_PAGES} pages — stopped early rather than return an incomplete list",
-                        self.id().as_str()
-                    )))
-                }
-            }
-        }
-        Ok(all)
+        self.list_models_transport(BASE, &api_key, timeouts::LIST_MODELS_TOTAL)
+            .await
     }
 
     async fn test_key(&self, app: &AppHandle) -> AppResult<()> {
